fix(contractor/contractor_status): Back to naive explanation generation
The current way of generating an explanation is not correct because a
witness is only responsible for the current local inconsistency (for a
given box) while we need to provide an explanation for the global
consistency (for the assertions).

We fall back to the naive approach which returns all C_i which have
contributed in the pruning process.