fix(contractor/contractor_status): Change unsat_witness to be Variables
Since we do not have an explicit formula for a variable to be an
integer, ContractorInteger does not have an associated formula with
it. We change the unsat_witness to be `Variables` instead of a vector
of `Formula`s to resolve the issue.

ContratorInteger is reporting the variable which reduces its
corresponding interval in the box to empty.