fix(solver): Remove is_model_variable property from symbolic::Variable
Motivation
----------

It is possible that a variable is used as a non-model variable in one
place and as a model-variable in another place. An example is a
exist-forall problem with a new encoding that introduces a new
variable `z` for `f(x)`. This variable `z` is used as a non-model
variable in the outer-most formulation (∃ part). But it should be used
as a model-variable in the counterexample generation phase (handling ∀).

Solution
--------

Instead, we add `mark_model_variable` and `is_model_variable` methods
in `Context::Impl` class.